Don't use 'tmp' dir when sharing a single file
The file paths are different when sharing a single file and whole
repository. While sharing a repository, Working Copy zip its contents
and store the .zip file in 'tmp' directory. As for sharing a single
file, Working Copy does not use 'tmp' directory. In this commit the
script checks if 'tmp' is in file path pattern to get the correct
basename.